backend/pyproject.toml
[project]
name = "doc-translation-backend"
version = "0.1.0"
description = "Backend API for Doc Translation System"
requires-python = ">=3.12"
dependencies = [
    "strawberry-graphql[asgi]>=0.287.0",
    "uvicorn[standard]>=0.38.0",
    "PyJWT>=2.10.1",
    "bcrypt>=5.0.0",
    "openpyxl>=3.1.5",
    "boto3>=1.41.2",
    "Pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "PyMuPDF>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "hypothesis>=6.148.2",
    "httpx>=0.28.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.setuptools]
packages = ["src"]
